Rename Crimegraphics' config file to CG_configs to prevent confusion
