+++ Base_Scripts/Scrapers/crimegraphics/CG_configs.py
